Deplacement polyline en AjoutTrajetActivity mais non fonctionnel
