+++ app/src/main/java/com/example/galidog2/LectureActivity.java
@@ -8,8 +8,6 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Paint;
 import android.graphics.drawable.Drawable;
-import android.location.LocationListener;
-import android.location.LocationManager;
 import android.os.Bundle;
 import android.os.Environment;
 import android.preference.PreferenceManager;
@@ -30,7 +28,6 @@ import org.osmdroid.views.MapView;
 import org.osmdroid.views.overlay.FolderOverlay;
 import org.osmdroid.views.overlay.MapEventsOverlay;
 import org.osmdroid.views.overlay.Overlay;
-import org.osmdroid.views.overlay.Polyline;
 import org.osmdroid.views.overlay.ScaleBarOverlay;
 import org.osmdroid.views.overlay.mylocation.GpsMyLocationProvider;
 import org.osmdroid.views.overlay.mylocation.MyLocationNewOverlay;
@@ -42,8 +39,6 @@ import org.osmdroid.views.overlay.simplefastpoint.SimplePointTheme;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Timer;
-import java.util.TimerTask;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
@@ -66,11 +61,8 @@ public class LectureActivity extends AppCompatActivity implements MapEventsRecei
     private int MY_PERMISSIONS_REQUEST_ACCESS_COARSE_LOCATION;
     private int MY_PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION;
     private String nomFichier;
-    private Polyline polyline;
     //Liste des points à marquer
     List<IGeoPoint> points = new ArrayList<>();
-    KmlDocument kmlDocument = new KmlDocument();
-
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -93,41 +85,8 @@ public class LectureActivity extends AppCompatActivity implements MapEventsRecei
 
         MapEventsOverlay mapEventsOverlay = new MapEventsOverlay(this);
         map.getOverlays().add(0, mapEventsOverlay);
-
-        tracerPolyline();
-
-//        arreterTrajet();
     }
 
-    LocationListener locationListener = new LocationListener() {
-        @Override
-        public void onLocationChanged(android.location.Location location) {
-            double latitude = location.getLatitude();
-            double longitude = location.getLongitude();
-            GeoPoint geoPoint = new GeoPoint(latitude, longitude);
-
-            String msg = "New Latitude: " + latitude + "\n New Longitude: " + longitude;
-            Toast.makeText(LectureActivity.this, msg, Toast.LENGTH_LONG).show();
-
-            polyline.addPoint(geoPoint);
-            map.getOverlays().add(polyline);
-        }
-
-        @Override
-        public void onStatusChanged(String provider, int status, Bundle extras) {
-
-        }
-
-        @Override
-        public void onProviderEnabled(String provider) {
-
-        }
-
-        @Override
-        public void onProviderDisabled(String provider) {
-
-        }
-    };
 
     /**
      * Fonction utilisée lorsque le mal-voyant refait seul la route
@@ -144,42 +103,7 @@ public class LectureActivity extends AppCompatActivity implements MapEventsRecei
     /**
      * Arret après 20 secondes pour essai
      */
-    private void arreterTrajet() {
-        long delay = 20 * 1000;
-        Timer timer = new Timer();
-        timer.schedule(new TimerTask() {
-            @Override
-            public void run() {
-                kmlDocument.mKmlRoot.addOverlay(polyline, kmlDocument);
-                File localFile = kmlDocument.getDefaultPathForAndroid("my_route.kml");
-                kmlDocument.saveAsKML(localFile);
-                Toast.makeText(LectureActivity.this, "Enregistré", Toast.LENGTH_SHORT).show();
-            }
-        }, delay);
-    }
 
-    private void tracerPolyline() {
-        polyline = new Polyline(map);
-        polyline.setWidth(8f);
-
-        int minTime = 4000;
-        int minDistance = 2;
-
-        LocationManager locationManager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
-
-        //Nécessaire pour pas d'erreur mais degueulasse !
-        if (checkSelfPermission(Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && checkSelfPermission(Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-            // TODO: Consider calling
-            //    Activity#requestPermissions
-            // here to request the missing permissions, and then overriding
-            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            //                                          int[] grantResults)
-            // to handle the case where the user grants the permission. See the documentation
-            // for Activity#requestPermissions for more details.
-            return;
-        }
-        locationManager.requestLocationUpdates(LocationManager.GPS_PROVIDER, minTime, minDistance, locationListener);
-    }
 
     private void demandePermissionsLocalisation() {
         if (ContextCompat.checkSelfPermission(this,
